refactor wip: move explorer admin routes to adminRouter
